Bring all code with unit tests up to 85% minimum coverage and enforce (#50)
* Simplify since unused args can simply be unnammed.

* Add full coverage of BuildServer and Serve

* Rename to mtls_test.go and add tests for all combos.

Make flags return an error instead of panic. If called from init it can panic.
This allows testing easier.

* Add build.go to generate binaries

* Finish out test coverage

* Add 2 more test helpers:

One when comparing error expected states (boolean).

General cmp.Diff

* Add FatalOnNoErr for other common error checking case

* Add op to WantErr and DiffErr

* Refactor to use new testutil functions and simplify some code

* Refactor build step to use build.go

Move paths to use bin/

Redo testing to force coverage and validate on paths we care about we have minimum coverage.

* Log response if we get no error when we thought we should

* Complete coverage for rest of rpcauth package

* Remove proxy/proxy from required coverage for now (no tests).

Leave TODO. Next CL

* Make sure we close the blob and the reader in all cases

* Add tsan checking separate from coverage

* Wrap srv inside of a mutex and helper funcs so tsan doesn't complain in tests

* Attempt to run integration tests

* Remove integration for now. needs work

* Embed errors with %w

* Update comments explaining transform is a noop for non-proto

+++ testing/integrate.sh
@@ -18,7 +18,9 @@ function check_status {
 
 function shutdown {
   echo "Shutting down"
-  kill -KILL ${PROXY_PID}
+  if [ -n "${PROXY_PID}" ]; then
+    kill -KILL ${PROXY_PID}
+  fi
   aws s3 rm s3://${USER}-dev/hosts
   sudo killall sansshell-server
 }
@@ -173,6 +175,8 @@ function check_perms_mode {
   fi
 }
 
+PROXY_PID=""
+
 OS=$(uname -s)
 if [ "${OS}" != "Linux" ]; then
   echo "Integration testing only works on linux"
@@ -191,21 +195,93 @@ cd $(dirname $PWD/$BASH_SOURCE)/..
 echo "package sansshell.authz" > ${LOGS}/policy
 echo "default allow = true" >> ${LOGS}/policy
 
-# Build everything (this won't rebuild the binaries)
+# Build everything (this won't rebuild the binaries but the generate will)
+echo
+echo "Running builds"
+echo
 go build -v ./...
-cd cmd/proxy-server
-go build -v
-cd ../sanssh
-go build -v
-cd ../sansshell-server
-go build -v 
-cd ../..
+go generate build.go
 
 check_status $? build
 
 # Test everything
-go test -v ./...
+echo
+echo "Running tests (with tsan)"
+echo
+go test -count=1 -race -timeout 30s  -v ./... 
 check_status $? test
+
+echo "Checking coverage - logs in ${LOGS}/cover.log"
+echo
+go test -timeout 30s -v -coverprofile=/tmp/go-code-cover 2>&1 | tee ${LOGS}/cover.log
+check_status $? coverage
+
+egrep ^ok.*coverage:.*of.statements\|'no test files' ${LOGS}/cover.log > ${LOGS}/cover-filtered.log
+
+# There are a bunch of directories where having no tests is fine.
+# They are either binaries, top level package directories or
+# testing code.
+# TODO(jchacon): Write these tests for proxy/proxy !
+for i in \
+  github.com/Snowflake-Labs/sansshell/proxy/proxy \
+  github.com/Snowflake-Labs/sansshell/auth/mtls/flags \
+  github.com/Snowflake-Labs/sansshell/cmd \
+  github.com/Snowflake-Labs/sansshell/cmd/proxy-server \
+  github.com/Snowflake-Labs/sansshell/cmd/sanssh \
+  github.com/Snowflake-Labs/sansshell/cmd/sansshell-server \
+  github.com/Snowflake-Labs/sansshell/proxy \
+  github.com/Snowflake-Labs/sansshell/proxy/protoc-gen-go-grpcproxy \
+  github.com/Snowflake-Labs/sansshell/proxy/testutil \
+  github.com/Snowflake-Labs/sansshell/services \
+  github.com/Snowflake-Labs/sansshell/services/ansible \
+  github.com/Snowflake-Labs/sansshell/services/ansible/client \
+  github.com/Snowflake-Labs/sansshell/services/exec \
+  github.com/Snowflake-Labs/sansshell/services/exec/client \
+  github.com/Snowflake-Labs/sansshell/services/healthcheck \
+  github.com/Snowflake-Labs/sansshell/services/healthcheck/client \
+  github.com/Snowflake-Labs/sansshell/services/localfile \
+  github.com/Snowflake-Labs/sansshell/services/localfile/client \
+  github.com/Snowflake-Labs/sansshell/services/packages \
+  github.com/Snowflake-Labs/sansshell/services/packages/client \
+  github.com/Snowflake-Labs/sansshell/services/process \
+  github.com/Snowflake-Labs/sansshell/services/process/client \
+  github.com/Snowflake-Labs/sansshell/testing/testutil; do
+  
+  egrep -E -v "$i[[:space:]]+.no test file" ${LOGS}/cover-filtered.log > ${LOGS}/cover-filtered2.log
+  cp ${LOGS}/cover-filtered2.log ${LOGS}/cover-filtered.log
+done
+
+# Abort if either required packaged have no tests or coverage is below 85%
+# for any package.
+abort_tests=""
+egrep 'no test files' ${LOGS}/cover-filtered.log
+if [ $? == 0 ]; then
+  abort_tests="Packages with no tests"
+fi
+echo
+
+bad=false
+oIFS=${IFS}
+IFS="
+"
+for i in $(egrep % ${LOGS}/cover-filtered.log); do
+  percent=$(echo $i | sed -e "s,.*\(coverage:.*\),\1," | awk '{print $2}' | sed -e 's:\%::')
+  # Have to use bc as expr can't handle floats...
+  if [ $(printf "$percent < 85.0\n" | bc) == "1" ]; then
+    echo "Coverage not high enough"
+    echo $i
+    echo
+    bad=true
+  fi
+done
+IFS=${oIFS}
+
+if [ "$bad" == "true" ]; then
+  abort_tests="${abort_tests}  Packages with coverage too low"
+fi
+if [ -n "${abort_tests}" ]; then
+  check_status 1 "${abort_tests}"
+fi
 
 # Remove zziplib so we can reinstall it
 echo "Removing zziplib package so install can put it back"
@@ -213,13 +289,13 @@ sudo yum remove -y zziplib
 
 echo
 echo "Starting servers. Logs in ${LOGS}"
-./cmd/proxy-server/proxy-server --policy-file=${LOGS}/policy --hostport=localhost:50043 >& ${LOGS}/proxy.log &
+./bin/proxy-server --policy-file=${LOGS}/policy --hostport=localhost:50043 >& ${LOGS}/proxy.log &
 PROXY_PID=$!
 # Since we're controlling lifetime the shell can ignore this (avoids useless termination messages).
 disown %%
 
 # The server needs to be root in order for package installation tests (and the nodes run this as root).
-sudo --preserve-env=AWS_ACCESS_KEY_ID,AWS_SECRET_ACCESS_KEY -b ./cmd/sansshell-server/sansshell-server --policy-file=${LOGS}/policy --hostport=localhost:50042 >& ${LOGS}/server.log
+sudo --preserve-env=AWS_ACCESS_KEY_ID,AWS_SECRET_ACCESS_KEY -b ./bin/sansshell-server --policy-file=${LOGS}/policy --hostport=localhost:50042 >& ${LOGS}/server.log
 
 # Make sure remote cloud works.
 # TODO(jchacon): Plumb a test account in via a flag instead of assuming this works.
@@ -237,7 +313,7 @@ else
   check_status $? Making s3 bucket
 fi
 
-SANSSH_NOPROXY="./cmd/sanssh/sanssh --timeout=120s"
+SANSSH_NOPROXY="./bin/sanssh --timeout=120s"
 SANSSH_PROXY="${SANSSH_NOPROXY} --proxy=localhost:50043"
 SINGLE_TARGET="--targets=localhost:50042"
 MULTI_TARGETS="--targets=localhost:50042,localhost:50042"
@@ -341,7 +417,7 @@ echo "uid, etc checks passed"
 
 run_a_test false 0 cp --overwrite --uid=${EXPECTED_NEW_UID} --gid=${EXPECTED_NEW_GID} --mode=${EXPECTED_NEW_MODE} ${LOGS}/hosts ${LOGS}/cp-hosts
 check_perms_mode ${LOGS}/cp-hosts
-echo cp text with bucket syntax
+echo cp test with bucket syntax
 aws s3 cp ${LOGS}/hosts s3://${USER}-dev/hosts
 run_a_test false 0 cp --overwrite --uid=${EXPECTED_NEW_UID} --gid=${EXPECTED_NEW_GID} --mode=${EXPECTED_NEW_MODE} --bucket=s3://${USER}-dev?region=us-west-2 hosts ${LOGS}/cp-hosts
 check_perms_mode ${LOGS}/cp-hosts
